eigencompute/requirements.txt
# Deep42 Agent - Production Requirements
# ========================================

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Google Gemini AI
google-generativeai>=0.8.3

# HTTP client
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0
